Convert EncoderTest::last_pts_ to a local variable
Convert the data member EncoderTest::last_pts_ to a local variable in
the EncoderTest::RunLoop() method. EncoderTest::last_pts_ is only used
in that method, and that method first sets EncoderTest::last_pts_ to 0
before using it. So EncoderTest::last_pts_ is effectively a local
variable in that method.

Note that DatarateTest, a subclass of EncoderTest, declares its own
last_pts_ data member and uses it to calculate the data rate. Apparently
DatarateTest::last_pts_ hides the same-named data member in the base
class EncoderTest. Although this is allowed by C++, this is very
confusing.

Ref: https://chromium-review.googlesource.com/c/webm/libvpx/+/3700777

Change-Id: I159c42e66ce6724dbdba1507d1ecbf023f9d06c3