rtc: Fix to chroma_check in variance partition
The uv_sad in the chroma_check is based on the
dst buffer but that is only set when the motion vector
(from int_pro_motion) is nonzero. Otherwise we should
use pre[0] buffer. Also move the chroma_check higher
up, before the partitioning is set.

Neutral stats (bdrat and speed) observed on
speed 7, 9 rtc_derf, and speed 9 rtc set.
For screen speed 10: ~2% speedup, ~1% bdrate loss.

Change-Id: I528e8463934202d6dd1f5f6fc26f37588cdebf4c