rtc-screen: Avoid golden ref check in variance_partition
Add check to only test golden as a reference in variance
partitioning, if y source_sad is above kLow.
For screen golden as an additional long-term reference is not
as useful, so this saves some computation with minimal
bdrate loss.

Stats change for speed 10 screen: neutal bdrate,
very small speed (< 0.5%).

Change-Id: I6092f5d4b42f38e8a701787b110ceaf8c6bd1dc8