row_mt enc: Delay top-right sync when intraBC is enabled
This patch introduces a larger top-right delay when intraBC is
enabled to ensure that sufficient number of superblocks have
finished encoding in the previous rows before encoding a block with
intraBC. This ensures that row-multithreaded encoding respects the
hardware constraints put on parallel processing of intraBC and thus
prevents any invalid access while building the predictor. Also
renamed some variables used in encoder row multithreading to
represent their actual purpose and re-enabled the unit test
ScreenContentToolsMultiThreadTestLarge.ScreenContentToolsTest as
the data race issue with thread sanitizer is resolved with this
change.

BUG=aomedia:3278

Change-Id: I042434c69d4d8db784a99e6d0ae6241f2b538b49