row_mt enc: Delay top-right sync when intraBC is enabled
This patch introduces an extra top-right delay when the intraBC
tool is enabled so as to ensure that a sufficient number of
superblocks have finished encoding in the previous rows before
starting to encode the current superblock. This ensures that
row-multithreaded encoding respects the hardware constraints on
parallel processing with the intraBC tool and thus prevents any
invalid access while building the predictor. Also, the unit test
ScreenContentToolsMultiThreadTestLarge.ScreenContentToolsTest is
re-introduced with thread sanitizer as the data race issue is
resolved with this change.

BUG=aomedia:3278

Change-Id: I042434c69d4d8db784a99e6d0ae6241f2b538b49