row_mt dec: Delay top-right sync when intraBC is enabled
This patch introduces an extra top-right delay when the intraBC
tool is enabled so as to ensure that a sufficient number of
superblocks have finished decoding in the previous rows before
starting to decode the current superblock. This ensures that
row-multithreaded decoding respects the hardware constraints on
parallel processing with the intraBC tool and thus prevents any
invalid access while building the predictor.

Change-Id: I544cfb11b0b1a9789cd56937713cc1bb05c663fb