RTC: Search GLOBALMV at the end of inter mode search
In realtime non-rd mode, modes with the same mv as the previously
searched modes are skipped. This causes an issue with both NEARESTMV and
GLOBALMV contains the zero motion vector.

Currently, GLOBALMV is searched prior to NEARESTMV. So when both
GLOBALMV and NEARESTMV are 0, GLOBALMV is always used in place of
NEARESTMV. However, in most situations, NEARESTMV should be preferred.
This is apparently in the scenario where there is a static background
with a moving foreground: if GLOBALMV is used for the background, this
would significantly increase the cost of coding the moving foreground.

This commit changes the behavior so that GLOBALMV is searched at the end
of single inter mode search so GLOBALMV is mostly used when NEARESTMV is
not 0. Furthermore, this commit adds a check so that  when
NEAREST/NEAR/NEWMV contains zeromv, GLOBALMV is automatically searched
with the currently available transform results.

Performance:
| SPD_SET | TESTSET | AVG_PSNR | OVR_PSNR |  SSIM   |  SPD  |
|---------|---------|----------|----------|---------|-------|
|    9    |   rtc   | -0.258%  | -0.252%  | -0.282% | -0.6% |
|    9    |rtc_derf | -0.734%  | -0.734%  | -0.576% | -0.6% |
|---------|---------|----------|----------|---------|-------|
|   10    |   rtc   | -1.035%  | -1.053%  | -1.211% | +0.0% |
|   10    |rtc_derf | -1.005%  | -0.929%  | -0.785% | -0.3% |

STATS_CHANGED

Change-Id: Id62ac0d448b66c3f6eb873ac4565490094fc679a