RTC: Search GLOBALMV after NEAREST/NEARMV
In realtime non-rd mode, modes with the same mv as the previously
searched modes are skipped. This causes an issue with both NEARESTMV and
GLOBALMV contains the zero motion vector.

Currently, GLOBALMV is searched prior to NEARESTMV. So when both
GLOBALMV and NEARESTMV are 0, GLOBALMV is always used in place of
NEARESTMV. However, in most situations, NEARESTMV should be preferred.
This is apparently in the scenario where there is a static background
with a moving foreground: if GLOBALMV is used for the background, this
would significantly increase the cost of coding the moving foreground.

This commit changes the behavior so that GLOBALMV is searched after
NEARESTMV and NEARMV, but before NEWMV. Furthermore, this commit adds a
check so that  when NEAREST/NEAR/NEWMV contains zeromv, GLOBALMV is
automatically searched with the currently available transform results.

Performance:
| SPD_SET | TESTSET  | AVG_PSNR | OVR_PSNR |  SSIM   |  SPD  |
|---------|----------|----------|----------|---------|-------|
|    9    |   rtc    | -0.229%  | -0.218%  | -0.207% | -0.5% |
|    9    | rtc_derf | -0.864%  | -0.857%  | -0.773% | -0.6% |
|---------|----------|----------|----------|---------|-------|
|   10    |   rtc    | -1.200%  | -1.196%  | -1.304% | +0.1% |
|   10    | rtc_derf | -1.217%  | -1.150%  | -1.118% | -0.2% |
|   10    |rtc_screen| -0.253%  | -0.393%  | -0.369% | +0.7% |

STATS_CHANGED

Change-Id: Id62ac0d448b66c3f6eb873ac4565490094fc679a