rtc: Fix uv sad calculation
The y_sad used in chroma_check for setting the
color_sensitivity is based on last frame so uv_sad
based on last frame should be used in the comparison.
However, for zero motion, uv_sad based on golden frame
was being used sometimes, as the predictor could be set
based on golden frame in the function setup_planes().
To overcome this issue, last frame predictor block
is setup in chroma_check function.

Also, in set_color_sensitivity() function, the predictor
buffer used in uv_sad calculation is corrected, as 'pd->dst'
is not setup before.

STATS_CHANGED

Change-Id: I50ac5a1492bf938da9897a6d3290c0c0242e06fa