rtc: Fix to disable segm for aq-mode=3
Disable the segmentation when the map is not set,
this case happen for content with high motion
(where cr->map may never be 0 on some frames).
Disabling and setting it to 0 reduces overhead and
gives small bdrate gain on a few clips (others have no change).

For speed 10 rtc_derf: apprtc, still_bright and street
get ~0.5% bdrate gain.

Change-Id: I5ddcf9654ed884d8d9682998424bf94c4061fbf5