Include Status in return values of all Q Mode API calls.
This is done by introducing a StatusOr<T> template, based on the one
in Abseil (but greatly simplified).

This is a breaking API change, which will require changes to downstream
repos when they are updated.

For now DetermineGopInfo and GetGopEncodeInfo never return a failure
status. Once downstream callers support the new API, they may bee changed
to report errors for invalid inputs.

Bug: b/234480857
Change-Id: Ia841a9f77664a2879bc6c41c5490477c4efa7096