rtc-svc: Speed up temporal layer encoding
Add logging to sample encoder to prinout
encode_time/fps per layer.

Generally we found that the 3TL encoder for AV1 vs VP9
has larger speed gap (excess of ~10%) than the single
layer encoding. Most of the difference seems to come from
the TL1 and TL2 layers (these layers much faster in vp9).

This CL keeps the larger motion search only for TL0
(as was intendent in the comments), so TL1 is now
faster/closer to vp9.

For TL2 (non-reference frames) at speed >= 9:
aaggressive_skip speed feature is enabled and cdf update
is disabled. There is a selective_mode to the
cdf_update control which will disable it for non-reference frames
for svc, but we prefer to add a speedup feature to force this
disabling on non-reference frames for speeds >= 9.

For 3TL lm=2 pattern (single reference) at speed 9:
avg bdrate loss of ~1.3% (max loss of ~2-3% on two clips).
Average speedup (encoder-fps) of ~10%.
This now makes the speed gap for 3TL similar to 1 layer.

Example command line:
examples/svc_encoder_rtc --speed=9 -t 1/30 --max-q=52 --spatial-layers=1
--temporal-layers=3 -lm 2 --threads=1 --aqmode=1 --target-bitrate=400
--bitrates=220,300,400 --error-resilient=0 input.y4m -o out.av1

Change-Id: I9496f9685bdc7f3e5b2402bf22c9d52ae71f5372