Fix issue 3296.
In the current implementation of the AV1 real-time mode,
when non rd pick mode is set (spd >= 7), the max allowed transform
size is 16x16.

This is in conflict with the command option
"--enable-tx-size-search=0".

We ignore the option if the non rd pick mode is in use.

Fixed: aomedia:3296

Unit tests are added to test there is no failure for all intra
encoding.

Different "--deltaq-mode" are also tested.

Change-Id: I1eb554af3cb28083e4166c18bb8ed566f206028f