Allintra: Extend sf dc_blk_pred_level to allintra speed>=6
The speed feature dc_blk_pred_level determines if a block should
be coded as transform skip or dc only block based on residual
statistics. In this CL, the speed feature is extended to allintra
speed>=6 by introducing intermediate levels to the speed feature
and predict_dc_levels to check if the block should be coded as
transform skip. The existing speed feature levels 1 and 2 become
levels 2 and 3 and the existing level 1 of predict_dc_levels
becomes level 2.

For AVIF still-image encode,

             Encode Time      BD-Rate Loss(%)
cpu-used     Reduction(%)     psnr       ssim
   6            1.012         0.0827     0.0936

STATS_CHANGED

Change-Id: I017175e10d730d2f2bdb9945b44ff8a24c003c86