rtc-svc: Update compound mode for svc
Remove the svc logic in nonrd pickmode for
compound selection: so now compound is only checked for
blocks > 16x16 for all modes (speed feature to allow it
for small blocks may be back added later).
And set the check_only_zero_zeromv_on_large_blocks for svc
in the speed features (false for speed <= 9, so keep
nearsest_nearest check on base for speed <= 9).

Stats for speed 9 3TL rtc set:
~2% speedup, 0.75% avg, bdrate loss

Two outliers clips desktop360p and tacomascmvvga are ~3-5% loss,
presumably this is due to removing compouund on small blocks,
will be looked into later.

Change-Id: If3ff7bafaeaeb255f7dc54c2259704992e588319