x86: normalize types used with _mm_cvtsi128_si32
prefer int in most cases to match the return type of the function; for
consistency intra predictors are kept unsigned as some implementations
rely on this in subsequent operations. this was the cause of previous
failures.

w/clang -fsanitize=integer fixes warnings of the form:
implicit conversion from type 'int' of value -809931979 (32-bit, signed)
to type 'uint32_t' (aka 'unsigned int') changed the value to 3485035317
(32-bit, unsigned)

Bug: b/229626362
Change-Id: I3c4442d9ab7263e8425de2be188b7c90ea3591e9