Use lane-referencing intrinsics in Neon convolution functions
The Neon convolution functions take a pointer to a filter and load
each of the (usually 8) values into separate Neon registers before
executing the series of multiply-accumulate instructions.

This patch modifies these helper functions to load all of the filter
values into a single Neon register and then access them via the lane-
referencing versions of the various multiply and multiply-accumulate
Neon instructions. This reduces register pressure and also the number
of load instructions.

Change-Id: I2c13251449113b26500517d9c6d774e189cd622e