Add recrf_rate and recrf_dist to Q Mode RC API
These fields are optional, but will either be available on all blocks
or on none (and likewise, should be present on all frames if they are
present on any).

The field TplFrameStats::rate_dist_present indicates whether the
fields in a frame are populated.

Also changed the width, height, row, and col fields to int16_t
to avoid increasing the memory footprint of the TPL stats any more
than necessary (b/241589513).
Before this CL, sizeof(TplBlockStats) = 64
After  this CL, sizeof(TplBlockStats) = 72

Bug: b/236983774
Change-Id: I29ad537a15887afa21846e31f7f5b889c60ea6f6