rtc: Fix to color artifacts under high motion
For RTC golden reference is used as longer-term
reference, but the color_sensitivity flags are
only defined relative to the last reference.
This will cause color artifacts for high motion
if golden is selected as the best reference and
has different uv-signal than source.

The fix here is to define a color_sensitivity for
golden reference at superblock level, and for now
skip testing the golden reference if this color
flag is set.

Bug:aomedia:3280

Change-Id: I17bdef4e14f416fb64d4ec49d269a90bcaec084a