rtc: Fix to color artifacts under high motion
For RTC: the golden reference is used as longer-term
reference, but the color_sensitivity flags are
only defined relative to the last reference.
This can cause color artifacts for high motion
if golden is selected as the best reference and
has different uv-signal than the source.

The fix here is to define a color_sensitivity for
golden reference at the superblock level, and for
now skip testing the golden reference if this
color flag is set. Logic is also based on source_variance
and content_stats, to avoid any regression from too
much/unneeded skipping.

This reduces color artifacts in the issue below.
To reduce further: logic above on golden-ref can be better
tuned, and defining color_sensitivity for alt-ref would
also be needed, but since alt-ref is closer to last ref,
the impact from that is less (so may be done in future CL).

Stats changed are neutral/small:
               avg-psnr  ovr-pnsr  ssim   IS-speedup
speed 7 derf:  -0.11     -0.12    -0.13    0.05
speed 9 derf:  -0.02     -0.00    -0.02   -0.17
speed 9 rtc:    0.08      0.07     0.04   -0.10
speed 10 rtc:  -0.07     -0.08    -0.03    0.40
speed 10 screen:0.26      0.09     0.34    0.04

Bug:aomedia:3308

Change-Id: I17bdef4e14f416fb64d4ec49d269a90bcaec084a