rtc: Fix to color artifacts under high motion
For RTC: the golden reference is used as longer-term
reference, but the color_sensitivity flags are
only defined relative to the last reference.
This can cause color artifacts for high motion
if golden is selected as the best reference and
has different uv-signal than the source.

The fix here is to define a color_sensitivity for
golden reference at the superblock level, and for
now skip testing the golden reference if this
color flag is set.

This reduces color artifacts in the issue below.

Bug:aomedia:3308

Change-Id: I17bdef4e14f416fb64d4ec49d269a90bcaec084a