Rename GetGopEncodeInfo in AV1RateControlQMode
AV1RateControlQMode now implements GetGopEncodeInfoWithLookahead
instead of GetGopEncodeInfo, but at this point it ignores the
lookahead_stats argument.

Also, GetGopEncodeInfo is no longer pure virtual, and the default
implementation returns an error. This will allow derived classes
to only implement GetGopEncodeInfoWithLookahead, so that remaining
references to GetGopEncodeInfo can be removed.

Bug: b/242892473
Change-Id: Iea2105e96f5601827139458029c48a0e185f9f0e