Real time rate control patch
This first part of this patch removes a cap on adaptation for overshoot
(suggested by MP) to reduce rate spikes and the tendency to overshoot
the target rate for some screen clips. This change has a net -ve impact
on metrics for the screen set of ~5% but little impact on the other
sets.

To balance this, the patch also includes new code to adjust the
refresh %  and refresh delta Q. This negates the metrics impact of the
first part of the patch on the screen set and even adds some further
metrics gains, but retains improvements in terms of rate spikes and rate
accuracy. It also gives significant net metrics gains for all of the
other test sets.

Metrics results re-based on current head  (opsnr, ssim)

speed 8
derf	-0.819	-1.802
rtc	-1.165	-1.859
1080	-2.512	-2.715
screen	-4.031	-3.859

speed 10
derf	-0.486	-1.355
rtc	-1.170	-1.944
1080	-4.967	-6.068
screen	 0.003	-0.564

Change-Id: Ifbb9b54f427a080751242b006a72088e531c2fb9