Real time rate control patch
This first part of this patch removes a cap on adaptation for overshoot
(suggested by MP) to reduce rate spikes and the tendency to overshoot
the target rate for some screen clips. This change has a net -ve impact
on metrics for the screen set of ~5% but little impact on the other
sets.

To balance this, the patch also includes new code to adjust the
refresh %  and refresh delta Q. This negates the metrics impact of the
first part of the patch on the screen set and even adds some further
metrics gains, but retains improvements in terms of rate spikes and rate
accuracy. It also gives significant net metrics gains for all of the
other test sets.

Metrics results (opsnr, ssim) -ve values better:

derf	-0.510	-1.251
rtc	-0.879	-1.644
1080p	-4.777	-5.975
screen	-5.320	-5.390

Change-Id: Ifbb9b54f427a080751242b006a72088e531c2fb9