Allocate tf_buf_second_arf buffer when required
The buffer 'tf_buf_second_arf' is used to hold the result of
temporal filtering corresponds to second arf frame. Currently,
the memory allocation is happening irrespective of the buffer
usage, which can lead to more memory consumption in some cases.
This CL allocates the memory for the 'tf_buf_second_arf' buffer,
only when it is required.

              Peak memory reduction(%)
Resolution    Single thread(good speed 6)
640x360             1.06
854x480             1.02
832x480             1.02
1280x720            1.10
1920x1080           1.09

HEAP memory reduction was measured using the following command.
$valgrind --tool=massif ./aomenc ...

Change-Id: I4d100a3205a307c620d8a9443d9406615af2aab7