RTC: Enable downsampled SAD for motion search
This commit modifies the use_downsampled_sad speed feature so that it
can be enabled at different block sizes. On 720p+, this is enabled on
block height \geq 16; on 480p, this is on when block height \geq 32. On
lower resolutions, this is turned off.

Performance:
| SPD_SET | TESTSET  | AVG_PSNR | OVR_PSNR |  SSIM   | ENC_T |
|---------|----------|----------|----------|---------|-------|
|    7    |   rtc    | -0.024%  | +0.002%  | +0.011% | -0.4% |
|---------|----------|----------|----------|---------|-------|
|    8    |   rtc    | +0.096%  | +0.081%  | +0.105% | -0.4% |
|---------|----------|----------|----------|---------|-------|
|    9    |   rtc    | +0.157%  | +0.141%  | +0.165% | -0.4% |
|---------|----------|----------|----------|---------|-------|
|   10    |   rtc    | +0.157%  | +0.163%  | +0.146% | -0.4% |

The clips 'thaloundeskmtgvga.y4m' and 'jimredvga_25fps.y4m' are
excluded from the metric calculation to reduce noise.

STATS_CHANGED

Change-Id: I9d1fd5318e94502d7790f616dc5f0178c8d72691