Constrain Q adjustment.
This patch modifies the constraints on Q adjustment between
frames for RTC.

Previously the maximum downward Q adjustment was 8 for
screen content and 16 for other content, This change makes
the maximum downward adjustment depend on the value of
Q for the last frame.

The change helps reduce large data spikes that can occur
when the frame to frame Q drops too fast, especially at
low Q.

For the rtc and derf sets there is a small negative impact on
metrics but for the 1080P set and screen set there are gains.

%BDRATE opsnr, ssim (-ve is better).

Speed 10
derf	 0.168	 0.453
rtc	 0.099	 0.265
1080p	-0.191	-0.439
Screen	-7.362	-6.013

Change-Id: I9d0679df12749b55eb92fccd9c65cc3ce128b289