rtc-screen: Improve logic of using golden when color is set
This is an extesion of:
https://aomedia-review.googlesource.com/c/aom/+/162105

Instead of skipping golden completely, which is ~5-7% bdrate
loss on youtube clip in rtc_screen, only skip if golden
is not set to be a better reference (based on sad in variance part).
If golden is set to be better referene than last
(i.e,ref_frame_prune = 0), then force the
color_sensitivity[0/1]=1 so UV channel is then factored
into the rd cost.

Change-Id: I2e5952cc654c98cfc8b71db6014ad29dca680671