rtc-screen: Improve logic of using golden when color is set
This is an extesion of:
https://aomedia-review.googlesource.com/c/aom/+/162105

Instead of skipping golden completely, which is ~5-7% bdrate
loss on youtube clip in rtc_screen, only skip if golden
is not set to be a better reference (based on sb sad in
variance part). If golden is set to be better reference
(i.e, nonrd_prune_ref_frame_search = 0), then force
color_sensitivity[0/1]=1 so color is then factored into rd cost.

This brings back ~5-7% bdrate gain on youtube clip,
otherwise neutral.
Change-Id: I2e5952cc654c98cfc8b71db6014ad29dca680671

Change-Id: I436db22364a2905c532fad5ec22d92ac0ede8311