Allintra: Avoid the application of loopfilters
This CL adds an API parameter 'apply_loopfilter' which can be used
to enable/disable the application of loopfilters. This parameter is
1 by default, which implies loop filtering should be applied.
Currently, the parameter is respected for ALLINTRA encoding only.
If loop filtering stages (cdef, superres scaling, loop restoration)
are disabled, the application of de-blocking is disabled.
This CL produces a bit-exact output stream.

Change-Id: Icfe91c3a485868e9e42f6816e8b766f812b56eea