Add ALTREF ref frame into ref_mode_set_reduced
For RTC speed >= 9, ref_mode_set_reduced is used to specify the testing modes and ref frames, but ALTREF is not tested by the configuration of current ref_mode_set_reduced.

This wasn't an issue in current speed feature setup because ALTREF as a single reference frame is turned off when speed >= 8, but it will potentially be an issue if someone wants to change the reference structure (i.e., turn on ALTREF as a single ref frame at speed >= 8).

Marco noticed this bug, and I tested and confirmed no behaviour change for speed = 7/8/9/10 based on our current reference structure setup.

Change-Id: Idd9f1bff936be8d6591574858d09b4f3a81044fe