rtc: Fix to 4x4avg in variance partition.
If 4x4avg is used in variance partition we need
to avoid the superblock boundary (by 4 pixels) if
temporal filtering is enabled. This is because the
4x4avg will load 8 pixels and temporal filter may
change source data, so when multi-threading is enabled
it will cause tsan race issue.

This is the cause of the issue in: Bug:aomedia:3331

No stats changed as 4x4avg is currently only used
on key frames, which does not have temporal filtering
or denoising.

Change-Id: Ie1df4c25c34924fd243bab935ffd232f9c89f4b4