Disable a speed feature that causes visual artifacts
The speed feature introduces visual artifacts reported in the buganizer.
BUG=b/245416468

The speed feature intends to prune NEARESTMV and NEARMV when the
count of spatial neighboring mvs is smaller than a threshold in
the ref mv stack.

However, the first coding block in a tile does not have neighboring mvs,
therefore it will always skip NEARESTMV and NEARMV mode.

For certain conditions where the mv difference is zero, NEWMV mode is
also early terminated. Therefore no inter mode is ever
evaludated in the rate-distortion optimization. This leads to the
selection of intra coding modes while an inter prediciton is obviously
a better trade-off. The selection of intra modes also leads to strong
visual artifacts since the prediction is very bad.

This phenomenon appear more frequenty for high resolution videos when
divided to sevral tile rows and columns.

We thus turn off the speed feature to allow the rate-distortion
optimization process to correctly choose inter modes, and artifacts
are gone consequently.

In addition, since it corrects the RD search, a better coding
performance is seen. On a 2K test video, the coding performance
before and after the fix is:
        ovr_psnr   bps
Before: 41.165    14174341
After:  41.381    14132105

Change-Id: Ie60061cd0d47b7cb1844d47e793e67bfa8ba1ad4