rtc: Extend sf reduce_mv_pel_precision_highmotion for cpu 7/8
This patch introduces a new conservative level ( = 1) for the
sf reduce_mv_pel_precision_highmotion. This level is tuned for
cpu 7 and 8 for HDRES content. For cpu 7, 8 HDRES max qty. drops
are 0.83% and 0.67% respectively, with no visual quality impact.

The Borg test results of RT preset:
 ______________________________________________________________
|cpu | Testset | Instr. Count |         BD-Rate Drop (%)       |
|    |         | Reduction(%) | avg. psnr | ovr.psnr |   ssim  |
|____|_________|______________|___________|__________|_________|
|  7 | HDRES   |    1.160     |  0.0901   |   0.1018 |  0.1482 |
|  8 | HDRES   |    0.605     | -0.0426   |  -0.0152 | -0.0014 |
|____|_________|______________|___________|__________|_________|

No changes to speed 9,10 and other testsets.

STATS_CHANGED

Change-Id: Ib45bea910b1cb099a684daaa8a7038bb9a3a6857