Specialize av1_convolve_2d_sr_vert_neon for 6-tap filters
av1_convolve_2d_sr_neon can be called with 4-, 6-, 8- and 12-tap
filters, with 4- and 6-tap filters being padded out with 0s to enable
re-use of the same 8-tap code path. This is inefficient as we end up
spending a lot of time multiplying by, and adding, 0 - especially
since this function is more commonly called with 4- or 6-tap filters
than 8-tap filters.

This patch adds an av1_convolve_2d_sr_vert_neon code path specialized
for 6-tap filters. This new path is used for both 4-tap and 6-tap
filters to reduce the amount of redundant work.

Change-Id: I073a1bfb71de7fd99ef384bee81e2d17e1af2ada