Allintra: Avoid the application of post-processing filters
This CL adds an encoder config parameter 'skip_postproc_filtering'
which can be used to skip the application of post-processing filters.
This parameter is 0 by default, which implies the post-processing
filters should be applied. Currently, the parameter is respected for
ALLINTRA encoding only. If subsequent post-processing filtering stages
(cdef, superres scaling, loop restoration) are disabled, the application
of de-blocking filters is disabled. This CL produces a bit-exact output
stream.

Change-Id: Icfe91c3a485868e9e42f6816e8b766f812b56eea