Optimize the initialization of input buffer for CDEF search
In parent version, the entire intermediate input buffer is
filled with value CDEF_VERY_LARGE(0x4000) in
av1_cdef_mse_calc_block(). However it is sufficient to fill
part of the buffer with above value, only when the boundary
pixels needed are outside the valid frame boundary. In this
CL, the default initialization is removed and the buffer is
filled with CDEF_VERY_LARGE when required. This CL produces
bit-exact output.

For real-time encode,
                 Instruction Count      BD-Rate Loss(%)
cpu   Testset      Reduction(%)    avg.psnr ovr.psnr  ssim
 7     rtc           1.910          0.0000   0.0000  0.0000
 7   rtc-screen      1.810          0.0000   0.0000  0.0000
 8     rtc           2.612          0.0000   0.0000  0.0000
 8   rtc-screen      2.004          0.0000   0.0000  0.0000

Change-Id: I6b88d8304ce0e2f4b35002ba167a5a64e6e4a46e