Optimize the initialization of input buffer for CDEF search
In parent version, the entire intermediate input buffer is
filled with 0x4000 in av1_cdef_mse_calc_block(). However
it is sufficient to fill part of the buffer with above
value, only when the neighboring pixels needed are outside
the valid frame boundary. In this CL, the default
initialization is removed and the buffer which is filled
with 0x4000, only when needed. This CL produces bit-exact
output.

For real-time encode,
                Instruction Count      BD-Rate Loss(%)
cpu  Resolution   Reduction(%)     avg.psnr ovr.psnr  ssim
 7     rtc          1.907           0.0000   0.0000  0.0000
 8     rtc          2.609           0.0000   0.0000  0.0000

Change-Id: I6b88d8304ce0e2f4b35002ba167a5a64e6e4a46e