rtc: Fix reference frame in variance partition
For variance-based partitioning: mainly for
RPS usage, when LAST may not be a reference.

(1) add check that LAST is actually selected as a
reference for prediction before computing y_sad and
using it for the variance calculations. This is always
true for the internal/default RTC coding, but may not
be true if user sets the reference structure and
switches off the LAST reference (e.g., RPS method).

(2) add check for using ALTREF as a reference for
the variance, similar to the check for GOLDEN.
This is needed for RPS, but will change stats for
default RTC when ALTREF is used as a single reference.

Only affects speed 7-rtc/rtc_derf and speed 8-rtc.
Some/small bdrate gain for rtc set, no speed loss.

Stats changed:
(avg/ovr/ssim, IC-speedup)
speed 7 rtc:
-0.2/-0.18/-0.17, -0.02

speed 8 rtc:
-0.22/-0.27/-032, 0.2

speed 7 rtc-derf:
-0.06/-0.09/-0.13, -0.09

Change-Id: Ib7e95fe423813382d8606e572f40c8f21463f15b