rtc: Fix reference frame in variance partition
For variance-based partitioning:

(1) add check that LAST is actually selected as a
reference for prediction before computing y_sad and
using it for the variance calculations. This is always
true for the internal/default  RTC coding, but may not
be true if user sets the reference structure and
switches off the LAST reference (e.g., RPS method).

(2) add check for using ALTREF as a reference for
the variance, similar to the check for GOLDEN.
This is needed for RPS, but will change stats for
default RTC when ALTREF is used as a single reference
(speed 7 and speed 8-rtc_derf).

Stats changed:

Change-Id: Ib7e95fe423813382d8606e572f40c8f21463f15b