Speed-up error calculation in av1_cdef_search()
In parent version, the error is calculated in 16-bit domain
even for 8-bit encoding and filtering functions are called
even though CDEF strength is zero. In this CL, the error is
calculated in 8-bit domain and filtering function calls are
avoided when CDEF strength is zero, selectively. This CL
produces bit-exact output.

               Instruction Count        BD-Rate Loss(%)
cpu Resolution   Reduction(%)    avg.psnr  ovr.psnr    ssim
 7        rtc      0.387          0.0000    0.0000    0.0000
 7  rtc_1080p      0.915          0.0000    0.0000    0.0000
 7 rtc_screen      2.334          0.0000    0.0000    0.0000
 8        rtc      0.420          0.0000    0.0000    0.0000
 8  rtc_1080p      0.593          0.0000    0.0000    0.0000
 8 rtc_screen      2.524          0.0000    0.0000    0.0000
 9  rtc_1080p      1.484          0.0000    0.0000    0.0000
10  rtc_1080p      1.511          0.0000    0.0000    0.0000

Change-Id: I12f21a06152ad4d123d462661458eac46c485f64