WIP : rtc : Avoid dry-run encode selectively
When nonrd_check_partition_merge_mode speed feature is
enabled, the best partition structure out of
PARTITION_NONE and PARTITION_SPLIT, with no further split,
is chosen based on block properties and QIndex. During
PARTITION_SPLIT, the dry-run encode of last sub-partition
is avoided in this CL as relevant data is not needed. This
CL produces bit-exact output.

Change-Id: Ic7d999d880ee9386d22750069203cff87c506bd8