rtc: Fix to RPS encoding: disable frame_id_numbers
Disable the frame_id_numbers_present_flag when
ref_frame_config is used to set the reference pattern:
this allows for continued decoding on a recovery
frame (predict from known available reference)
after previous frames are lost/dropped (before decoder).
This applies to RPS, and is similar to dropping enhancement
layers frames for SVC.

As such the condition to disable frame_id_numbers is
modified to be (use_svc || rtc_ref->set_ref_frame_config).
Usage of set_ref_frame_config is required for the RPS method.

Added a unittest for RPS, for which this fix is needed
in order for test to pass.

Change-Id: Ief765f7ab6766262e94ef08710b20786d24a6a9b