rtc: Extend sf use_adaptive_subpel_search for cpu 9
This CL extends the sf use_adaptive_subpel_search to cpu 9
which conditionally downgrades the subpel search logic. A
new level is created to tune this for cpu 9 using additional
block properties such as source_var and QP.

The Borg test results are as follows:
 --------------------------------------------------------------
|cpu | Testset | Instr. Count |        BD-Rate Drop (%)        |
|    |         | Reduction(%) | avg. psnr | ovr.psnr |  ssim   |
|----|---------|--------------|-----------|----------|---------
|  9 |   rtc   |    1.139     |   0.2245  |  0.1940  |  0.2511 |
 --------------------------------------------------------------

Speed 10 and testsets rtc_derf, rtc_screen are not affected.

STATS_CHANGED for rt

Change-Id: I42ffa61c8c26393e0c6ecaeb6756079567dafe7a