rtc: Allow for error_resilient = 0/off for RPS
For RPS (reference picture selection):
on recovery frames, where LAST is not a reference,
set the primary_ref_frame to GOLDEN or ALTIREF
(whichever is selected as reference). This allows for
continued decoding (when some frames are lost/droppped)
even with error_reslient=0/off.

Set error_resilient=0/off in the RPS unittest to verify.

Change-Id: Ibc5607fd59e6d1a1725648540adcb3249b2c5c24