Specialize dist_wtd_convolve_2d_vert_neon for 6-tap filters
av1_dist_wtd_convolve_2d_neon can be called with 4-, 6- and 8-tap
filters, with 4- and 6-tap filters being padded out with 0s to enable
re-use of the same 8-tap code path. This is inefficient as we end up
spending a lot of time multiplying by, and adding, 0 - especially
since the most common filter size used with this function is 6-taps.

This patch adds a dist_wtd_convolve_2d_vert_neon code path
specialized for 6-tap filters. This new path is used for both 4-tap
and 6-tap filters to reduce the amount of redundant work.

Change-Id: I89912d28d33ab2d41728ee88d6ae73a8328a9634