ALLINTRA: Skip extend_frame_borders()
In this CL, the call to extend_frame_borders() is skipped
for ALLINTRA encoding. As recon frames are not referenced
in case of ALLINTRA encoding, extension of frame borders
is not required. This CL also improves the documentation
of should_skip_postproc_filtering() and adds a cosmetic
change around the call to loopfilter_frame() function. This
is a bit-exact change.

Change-Id: Ib964f2e66de5e90cc5722671490151ad77d18cea