rtc : Fix the sf skip_cdef_sb when SB size is 128x128
In parent version, the decision to skip CDEF filtering is
done and stored at 64x64 block level. This is consistent
with signaling in the bitstream. However after processing
the superblock, the skip decision of first 64x64 overwrites
the decisions of remaining 64x64 blocks which is not correct
when superblock size is 128x128. In this CL, the relevant
code is cleaned-up.

For rtc-screen,
     Instruction Count        BD-Rate Loss(%)
cpu     Reduction(%)    avg.psnr   ovr.psnr    ssim
9         -0.109        -0.8515    -0.8704   -0.9524
10        -0.194        -0.6478    -0.5520   -0.6602

STATS_CHANGED for --rt usage

Change-Id: I19de4a6aa3a3864feb18bcbc05a5476528c29e71