rtc: Fix to screen with temporal layers
The scene_detection and source_metrics_sb use
current source vs last_source (source of previous frame),
but for temporal layers the previous frames does not
always correspond to the prediction reference. The base
temporal layers (TL0) always predicts from the previous
base TL0, which is 2/4 frames away for #TL=2/3.

This CL fixes this by allocating buffer to keep track
of the source frame for base TL0, and use that for
last_source for lower temporal layers.

For screen: the scene_detection and source_metrics_sb have
big impact on quality and so this wrong choice of
last_source can have big compression efficiency loss.

Bug: aomedia:3346
Change-Id: Ic809feb19b87d0881c95b2749968f9e7931f22f8