rtc: Fix to screen with temporal layers
In real-time mode the scene_detection and source_metrics_sb
use current source vs last_source (source of previous frame),
but for temporal layers the previous frames does not
always correspond to the prediction reference. For example,
the base temporal layers (TL0) always predicts from the
previous base TL0, which is 2/4 frames away for #TL=2/3.

For screen the scene_detection and source_metrics_sb have
big impact on quality and so this wrong choice of
last_source for these features can have big loss in
compression efficiency.

This CL fixes this by keeping track of the source frame for
the LAST reference, and using that for the last_source.

Bug: aomedia:3346
Change-Id: Ic809feb19b87d0881c95b2749968f9e7931f22f8