rtc: Fix to screen with temporal layers
In real-time mode there are speed features, i.e.,
scene_detection and source_metrics_sb, that use
current source vs last_source (source of previous frame),
but for temporal layers the previous frames does not
always correspond to the prediction reference (LAST).
For example, the base temporal layers (TL0) always predicts
from the previous base TL0, which is 2/4 frames away for
number_temporal_layers = 2/3.

For screen: the scene_detection and source_metrics_sb have
big impact on quality and so this wrong choice of
last_source for these features can have big loss in
compression efficiency.

This CL fixes this by keeping track of the source frame for
the LAST reference, and using that for the last_source.

This fix only applies to number_spatial_layers = 1,
follow-up will be made to handle spatial layers.

Added screen option to svc_encoder_rtc.c, via the patch
attached in the issue below.

Bug: aomedia:3346
Change-Id: Ic809feb19b87d0881c95b2749968f9e7931f22f8