RTC: Treat scene changes more like key frames.
For RTC coding scene changes cause a reset of the Q
decision process and cyclic refresh but some aspects
of this are tied to  key frames and a scene change
does not necessarily trigger an intra only frame.

This patch causes the count of frames since last
key frame to be reset when there is a scene change
even if it is not coded as an intra only frame.

For the rtc, derf and 1080P test sets set the metrics
results are broadly neutral as these sets do not contain
multiple scene changes so any impact is really only on
the first few frames, but for the the screen content set
the gains are large.

Speed 8
derf	-0.048	-0.020
rtc	-0.062	-0.089
1080p	 0.112	 0.135
screen	-8.101	-6.897

Speed 10
derf	-0.012	 0.025
rtc	 0.060	 0.040
1080p	-0.347	-0.050
screen	-6.515	-6.081

STATS_CHANGED

Change-Id: If86df588ef82772da9e060d78a343fa1afcb08c6