Round in symbol cost LUT instead of indices
Center entries around i+0.5 instead of i.
Rounding the indices adds to complexity from the by requiring an
addition and a min. In practice, an entire clamp is removed by inlining
get_prob into av1_cost_symbol and relying on the range of the input.

Stats changed, will update with results and add the special tag thingy.

Change-Id: I856bc5dba326003e8cd6f66e5a612736fd13aa38