RT: Revive a speed feature to try split partition for nonrd mode
In nonrd mode, partition is determined by variance based partitioning.
Here, we ask the encoder to try split partition when none partition is
picked by variance based partitioning, and select the one with
smaller rdcost as encode decision.

This achieves a better coding performance with the sacrifice of speed.

It is guarded by the speed feature: nonrd_check_partition_split
By default it is turned off.

Change-Id: Ibe232c7fada57b6f4d82956cb529937f297f28f0