rtc: Fix and tune the sf cb_pred_filter_search
This sf skips the interp filter-search for a block (in
checkerboard pattern) and predict a suitable filter based on
top/left neighbors. This CL fixes the logic by disabling
the prediction (search is done) in the following cases:
1. If either neighbor is unavailable
2. If either neighbor is non-inter mode
3. If top and left neighbors have chosen different filters
e.g. top: EIGHTTAP_REGULAR left: EIGHTTAP_SMOOTH
4. For lower speed-presets (8,9) if both are EIGHTTAP_SMOOTH
Earlier (when enabled), it would select EIGHTTAP_REGULAR in
cases 1-3.

The CL also improves the sf logic by disabling it for blocks
with <= kVeryLowSad, where subpel needs to be more accurate
via the filter search.

The Borg test results are:

            Instruction Count        BD-Rate Loss(%)
cpu Testset   Reduction(%)    avg.psnr   ovr.psnr    ssim
 8    rtc      0.878           0.0622     0.0630    0.0141
 9    rtc      1.084           0.1934     0.1780    0.1420
10    rtc      1.082           0.2609     0.2418    0.3699

Based on the results, this sf is disabled for 1080p.

No changes to cpu 7 and rtc, rtc-screen testsets.

STATS_CHANGED for rt

Change-Id: I044c95d6ec3cf30dda9424eba8569da21f5cbbef