rtc: Remove qp_thresh in encodedframe_overshoot_cbr()
When a frames is detected as having high source sad
(rc->high_source_sad = 1, e.g., scene/slide change),
the QP is set to max-q and rate control is reset.
But this max-q and reset was only done if Q was lower
than some threshold (close to worst_quality).
Not much reason to have this qp threshold, and removing
it makes code cleaner and more consistent.

This is small/no effect on metrics, for screen speed 10:
avg_psnr/ovr_psnr/ssim, IC speedup
-0.09/0.05/-0.03, 0.005

Change-Id: Ia7e4e8d9be30171cf9931bb2cddd442e29f0ceaa