rtc-svc: Fixes for screen with spatial layers
Several fixes to make screen work well with spatial layers:

-Last_source is set based on superframe count to ensure
last_source is the previous source for all spatial layers
-Scene detection is only done on SLO with full resolution,
and the state of the detection is passed on to the other
spatial layers.
-Source_sad is enabled now for every spatial layer.
-Cyclic refresh counter on distance to scene change/max_q
is added to the layer context and updated per spatial layer.
-Allow prediction from lower spatial layer in nonrd_pickmode.

Unittest added for screen with spatial layers.

Bug: aomedia:3346

Change-Id: I87961ead6154fc3642071ab67e99a4595565d00d