rtc-svc: Fixes for screen with spatial layers
Several fixes to make screen work with spatial layers:

-last_source is set based on superframe count to ensure
last_source is the previous source for all spatial layers
-source_last_TL0 is used to ensure last_source corresponds
to the LAST reference.
-scene detection is only done on SLO with full resolution,
and the state of the detection is passed on to the other
spatial layers.
-source_sad is allowed for every spatial layer.
-cyclic refresh counter on distance to scene change/max_q
is added to the layer context and updated per spatial layer.
-allow for prediction from lower spatial layer in nonrd_pickmode.

Unittest added for screen with spatial layers.

No change in stats for default 1 layer encoding.

Bug: aomedia:3346

Change-Id: I87961ead6154fc3642071ab67e99a4595565d00d