rtc: Fix a potential bug related to building inter pred
This CL fixes building inter-predictor when compound mode
is enabled and reuse_inter_pred is disabled. This issue
was introduced by an earlier CL:
https://aomedia-review.googlesource.com/c/aom/+/165301

The change is bit-exact, since the existing sf's do not hit
this. However, this fix is needed for consistency and
extending the sf: cb_pred_filter_search to cpu 8.

Change-Id: I9f17dbbd2fbf593a8c870c4e6e575a9458e12316