Fix normalization in dq3 mode
Make the normalization calculation process account for the fact
that deltaq mode is operated in units of superblocks. This resolves
a mismatch between the normalization and superblock QP decision
process that leads to universally increased operating bit-rate
at low QPs.

For a few 4K test clips at low QPs, this improves the encoding
speed by 20-40% on a single threads.

BUG=b/257074966

Change-Id: I392ffcae9aeef40d63a2e610167043ab5899160a