rtc-screen: Fix Q setting for forced key frames
If the user in RTC sets/forces a key frame in the
stream the q for that frame will be the last_boosted_qindex
which can be very low, especially for screen, and cause very
large encoded key frame size.

This logic may be ok for low-resolution video, but can
cause excessive frames sizes for screen, so remove it for
screen mode for now. This reduces the key frame size by
~3-20x (on slidechanges clip, depending on the scene change);
quality is reduced, but gets cleaned up quickly with aq-mode=3.

No change in stats for rtc sets (no forced key frames by default).

Change-Id: I606cfe617698e09859231b0bb80b5b7ae212737e