rtc-screen: Fix Q setting for forced key frames
If the user in RTC sets/forces a key frame in the
stream the q for that frame will be the last_boosted_qindex
which can be very low, especially for screen, and cause very
large encoded key frame size.

This logic may be ok for low-resolution video, so remove it
here for now only for screen. This reduces the key frame size by
~5-10x  (on slidechanges clip), quality is reduced by cleaned
up some soon after.

Change-Id: I606cfe617698e09859231b0bb80b5b7ae212737e