Change to rate correction factor on scene change.
Some adjustments to the rate correction factor after a scene change
were still using assumptions from VP9.

Corrections to the calculation of the bits per pixel enumerator.
Slightly more conservative when in resetting correction factor.

The direct impact of this change on metrics in isolation is very small
for the derf ,rtc and 1080P sets. Modest gains for the screen set.

STATS_CHANGED

Change-Id: Id7dad33a66e0b9798e84dbd0995cabecfc6b0d4d