Don't update rcf on scene change.
When a cbr scene change is detected various values including the
rate control factor get adjusted or set to defaults in. Hence we
should not adjust them again here.

This patch, when added to the previous change to the adjustment
of the rate control factor in av1_encodedframe_overshoot_cbr()
gives a large gain in the screen set where scene changes are common
but has only a modest impact in the other sets.

Screen set speed 10 (opsnr, ssim)
-6.896	-4.842

STATS_CHANGED

Change-Id: I1f75332472801afcca7828fc1d570f9ef19ceca7