rtc-svc: Simplify conditions to skip application of LPF and CDEF
In the parent version, the application of deblocking and CDEF
filters is skipped in svc real-time encoding mode based on
cpi->ppi->rtc_ref.non_reference_frame. In this CL, the bitmask
skip_apply_postproc_filters is populated appropriately when
cpi->ppi->rtc_ref.non_reference_frame=1 and the bitmask is used
to avoid the application of deblocking and CDEF filters. This is
a bit-exact change.

Change-Id: Id08a71fca9377574c19d8c39f2b950cebee746b3