Vectorize disflow code
Most of the runtime in the disflow code is concentrated in 2
very hot functions and 2 more slightly less hot functions.
This commit vectorizes all of these functions.

In order to minimize overhead due to branching on CPU features, we
branch at the call to compute_flow_at_point(), which is the direct
caller of all of these hot functions.

The disflow output is the same before and after this commit,
so there is no BDRATE change. But there is a significant speed
improvement:

[TODO measure]

Change-Id: I162ca971480d97d4585aeead8b2750f0af44724f