rtc: Remove qp_thresh in encodedframe_overshoot_cbr()
When a frames is detected as having high source sad
(rc->high_source_sad = 1, e.g., scene/slide change),
the QP is set to max-q and rate control is reset.
But this max-q and reset was only done if Q was lower
than some threshold (close to worst_quality).
Not much reason to have this qp threshold, and removing
it makes code cleaner and more consistent.

This is small/no effect on metrics, for screen speed 10:
avg_psnr/ovr_psnr/ssim, IC speedup:

Change-Id: I64c9edc62b129f803fd5e9be1a3d7f1e62cb0284