svc-screen: Fix expectation in screen unittest
With temporal layers: the number of mismatch frames
is equal to the #non-reference frames (since encoder
skips loopfilter but decoder applies it). But for some
screen content the loopfilter levels may be so small
that the loopfilter has no effect on the decoder size,
in which case that frame won't be a mismatch.

So for temporal layers screen we should have:
number of mismatch <= #non-reference frames.

This was triggered by this CL:
https://aomedia-review.googlesource.com/c/aom/+/166461

Change-Id: Iac73d9618137961901ecb20b23591cef1a7e6edc