rtc-svc: Scale source_sad thresholds in cyclic_refresh
The source_sad is used to guide the refresh map,
but for temporal layers, where cyclic refresh is
only done on the base TL0, the source_sad (between
current and LAST source) is 2/4 frames away, and so
can be much larger. Here we increase the thresholds
accordingly to account for this.

bdrate gain for speed 10 3TL rtc_derf:

Change-Id: I9281a44b01f01ca330d5bc0a9e66eb7a1d80645b