Fix several bugs in K-means.
Mostly, a test was added to make sure total_dist is the same on all
platforms

Some introduced by the recent 8cbaea62af09e92a81853e1787c9912b3ac6999c:
- dist instead of *dist: the C version was not part of my SSIM check
- the horizontal_sum_avx2 function has been re-written to be pure SSE2
(to avoid non-aligned memory problem)
- int64_t conversions were wrong (though the results right because
only a comparison is needed and the results were shifted by 32 bits)

Some much older:
- av1_calc_indices_dim2_sse2 was actually wrong as it was doing
_mm_madd_epi16 on 32 bit which only works for positive numbers, and
is otherwise creating a difference of 1 (-1 * -1).
The solution is to take the absolute value for now. The proper
solution is to switch the pipeline to 16 bit as this is what it
assumes.

Change-Id: Iaeeefa9f41d5262a622f6b64b9221696a2335c8d