Modify encoder's entropy coding byte-push mechanism
Changed the byte-push mechanism in libaom entropy
encoder. In the parent version, the register is
flushed when at least 1 byte is available. In the
CL, we make the register 64-bit and flush it when
at least 4 bytes are accumulated in it.

No changes to the decoder logic. This still remains
as a To Do item.

<Update speed numbers>

Change-Id: If23e711145aef1fa193523c98b29c7c82eff76b7