Modify encoder's entropy coding byte-push mechanism
Changed the byte-push mechanism in libaom entropy
encoder. In the parent version, the register is
flushed when at least 1 byte is available. In the
CL, we make the register 64-bit and flush it when
at least 4 bytes are accumulated in it.

An intermediate 16-bit buffer 'precarry_buf' is
removed and carry-propagation mechanism is changed.

No changes to the decoder logic.

<Update speed numbers>

Change-Id: If23e711145aef1fa193523c98b29c7c82eff76b7