rtc: Fix crash in variance partition for resizing
On resized frame, or if LAST is NULL, the frame is treated
as key frame in variance partitioning (lines ~1406),
but in fill_variance_tree_leaves() the is_key_frame is
set again and goes not include logic from lines ~1406.

This causes a crash for resizing with 10bit input.

Fix is to pass the is_key_frame into
fill_variance_tree_leaves().

Change-Id: Ibe4162de54cf31739667796a9666880293bd54ad