Change the license to the BSD 2 Clause License
The BSD 3-Clause Clear License was used when these files were added or
moved in the following two CLs:
https://aomedia-review.googlesource.com/c/aom/+/167361
https://aomedia-review.googlesource.com/c/aom/+/166964

Until we have relicensed libaom under the BSD 3-Clause Clear License, it
would be better to avoid a mixture of licenses.

Also change the copyright year of the files (except aom_dsp/rect.h) back
to 2016 as the code in those files was moved from elsewhere in the
repository which said 2016.

Change-Id: I5bf0af13d338a86c6b1e448b0a12101d3fceda81