Avoid an unnecessary allocation
Force the encoder-side enable_global_motion flag to 0 in
realtime mode, as global motion is never used in this case.
This prevents us from allocating a buffer that we will never use.

Also change the memory accounting logic so that this buffer is
only accounted toward the maximum allocation limit if we are
actually going to allocate it.

Finally, add an assertion to ensure that this buffer is never
allocated in decoder-only or realtime-only builds.

Change-Id: I83a25fcd3bc98091fabe1caf79928e565a5e51a6