Avoid an unnecessary allocation
Force the encoder-side enable_global_motion flag to 0 in
realtime mode, as global motion is never used in this case.

This prevents an edge case where we could allocate a buffer
to store an 8-bit version of the source frame, but then not
actually use it. This could happen if using realtime mode
with 16-bit internal buffers, on a build with CONFIG_REALTIME_ONLY=0

Also change the memory accounting logic so that this buffer is
only accounted toward the maximum allocation limit if it is
actually allocated.

Finally, add an assertion to ensure that this buffer is never
allocated in decoder-only or realtime-only builds.

Change-Id: I83a25fcd3bc98091fabe1caf79928e565a5e51a6