Change param of AOME_SET_SPATIAL_LAYER_ID to int
Change the parameter type of AOME_SET_SPATIAL_LAYER_ID from unsigned int
to int. This change is binary compatible because the valid spatial layer
id values 0..3 can be represented in either the unsigned int or int
type. libaom uses the int type for spatial layer ids internally. So this
CL changes the API to match the implementation.

Change-Id: I585808d928ff106a8ae23e1993ba08f8020988e4