Avoid redundant calls to av1_build_quantizer()
The function av1_build_quantizer() initializes
different quantizer parameters for entire QINDEX_RANGE.
The quantizer tables need to be re-initialized only when the
dc/ac delta quantizer indices of any of the planes are modified.
This CL avoids the redundant calls to av1_build_quantizer() by
storing dc/ac delta quantizer indices for which the quantizer
tables were initialized before. This is a bit-exact change.

For AVIF still-image encode, an average encode time reduction of
~1.12% is observed on average across different resolutions with
speed=9 and threads=4.

Change-Id: Iba1f2108e3efaaec06ab2eb96b2f5561d18aa530