Cache corner list for each frame buffer
Store the corner list for each frame in a similar way to the
downsampling pyramid. This way, it can be computed once on-demand,
and then cached for later usage.

Now that both the pyramid and the corner list are cached, it becomes
possible to push the calls to compute these values down into the
per-estimation-type functions. This ensures that we only compute exactly
what we need, and the caching ensures that we don't do any redundant
calculations.

Finally, move the invalidation of cached global motion state in
encoder.c. Previously this logic was only in encode_with_recode_loop,
which meant that stale data could be used we used encode_without_recode
instead. By moving the logic up to a common ancestor function, we can
avoid this problem without duplicating code.

Results (150 frames, speed 1):

lowres2: 0.000% BDRATE, -0.124% encode time
midres2: 0.000% BDRATE, -0.118% encode time
hdres2:  0.000% BDRATE, -0.088% encode time

Change-Id: I473adebb63451cd91e0ec637e6d339037d130961