Cache corner list for each frame buffer
Store the corner list for each frame in a similar way to
the downsampling pyramid. This way, it can be computed once
on-demand and then cached for later usage.

Now that both the pyramid and the corner list are cached, it
becomes possible to push the calls to compute these values down
into the per-estimation-type functions. This ensures that we only
compute exactly the values we need, without the runtime cost
of recomputing them each time.

Results pending

Change-Id: I473adebb63451cd91e0ec637e6d339037d130961