av1_alloc_pc_tree_node: fix allocation check
Previously the code would use an uninitialized error variable which
could cause the code to jump to an arbitrary location on failure. The
code will still crash with this change; the allocation failure needs to
be propagated. This is left to a follow up given the call sites may be
in different threads and not have a way to exit gracefully currently.

Bug: aomedia:3276
Change-Id: Ib5d6e0ebdde563e465f92eea4218d9515ae3b2e1