Re-enable global motion + superres
The linked bug references a crash when combining global motion and superres.
The underlying issue was that av1_superres_post_encode() did not allocate
a necessary buffer for global motion estimation. Therefore, if a frame was
coded using superres and then used as a reference by a full-size frame,
the encoder would segfault when trying to do global motion estimation.

This was temporarily avoided by patch 5d99c65e by disabling global motion
when superres is enabled.

However, the patch before this fixes the underlying problem, so we can now
remove this temporary fix. This restores the option of using global motion
for frames which are coded at full size, even if other frames within the
same sequence use superres.

BUG=b/261325007

Change-Id: I71af86d668d52071d3342f94388e76bd15e8b494