Allintra: Reduce the instances of RefCntBuffer
The buffer frame_bufs of type RefCntBuffer is a member of
BufferPool and the same is used to hold information regarding the
reference frames. During allintra video encode, only a maximum of
two reference frames are required at any time whereas there were
8 instances of RefCntBuffer in the parent version. This CL avoids
the memory allocation of the instances of RefCntBuffer that are
not required for allintra video encode. This is a bit-exact change.

For AVIF image encode with speed = 9,

             HEAP Memory reduction(%)
Resolution   threads=1    threads=4
640x360         4.748       3.765
768x512         2.990       2.331
832x480         3.110       2.606
1280x720        1.466       1.331
4032x3024       0.128       0.126

For AVIF still-image encode, an average encode time reduction of
~0.68% is observed on average across different resolutions with
speed=9 and threads=4.

Change-Id: Id7d485b9c15a938f5951ebac6522cd01c47166cb