Introduce sf prune_rect_part all block sizes
In this CL, we introduced a speed feature,
prune_rect_part_based_on_4x4_var_deviation to prune rectangular
partitions for blocks of size 16X16 and above based on variance
of 4x4 sub-blocks in the block. If the difference between maximum
and minimum of sub-block variances is lesser, rectangular
partitions are pruned. This change is applicable only for
--allintra speed=6.

For AVIF still-image encode,

             Encode Time      BD-Rate Loss(%)
cpu-used     Reduction(%)      psnr       ssim
    6           8.100         0.1497     0.3790

STATS_CHANGED for speed=6

Change-Id: I3c88a52947408d0f3dab3c260b07483f2969cde1