Introduce sf prune_rect_part_using_4x4_var_deviation
In this CL, a speed feature,
prune_rect_part_using_4x4_var_deviation, is introduced
to prune rectangular partitions based on variance of 4x4
sub-blocks in the block. If the difference between
maximum and minimum of sub-block variances is lesser,
rectangular partitions are pruned. This change is
applicable only for --allintra speed=6.

For AVIF still-image encode,

             Encode Time      BD-Rate Loss(%)
cpu-used     Reduction(%)      psnr       ssim
    6           8.140         0.1539     0.3797

STATS_CHANGED for speed=6

Change-Id: I42f5f972d982eac2f0b24c1f071ab5b337522ac9